requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.1
django-cors-headers==1.0.0
django-toolbelt==0.0.1
djangorestframework==3.1.2
gunicorn==19.3.0
psycopg2==2.6
static3==0.6.0
